Fix bug in Edit Patron and Delete Patron command
The updated information of patrons are not reflected on the books that
they borrow, or requested for, when patron information is edited or
deleted. Furthremore, patrons that has unreturned books can be deleted
from the system.

Let's:
* Disallow the deletion of patrons who have unreturned books
* Update information of requesters of related books whenever a requester
patron is deleted.
* UPdate information of borrowers and requesters of related books
whenever a related patron is edited.

Disallowing the deletion of patrons who have unreturned books will
ensure that books can only be made available throught Add Book and
Return Book command, therefore helps with data integrity. Making a book
avaialble through delete patron command would mean that all requesters
for books with the same isbn will need to be deleted as well, resulting
in data inconsistency if not handled properly.

+++ src/main/java/seedu/address/logic/commands/patron/DeletePatronCommand.java
@@ -22,7 +22,8 @@ public class DeletePatronCommand extends Command {
             + "Parameters: INDEX (must be a positive integer)\n"
             + "Example: " + PATRON_COMMAND_GROUP + DELETE_COMMAND_WORD + " 1";
 
-    public static final String MESSAGE_DELETE_PATRON_SUCCESS = "Deleted Patron: %1$s";
+    public static final String MESSAGE_DELETE_PATRON_SUCCESS = "Deleted Patron: %1$s\n";
+    public static final String MESSAGE_HAS_BORROWED_BOOKS = "You cannot delete a patron that has unreturned books!";
 
     private final Index targetIndex;
 
@@ -40,8 +41,15 @@ public class DeletePatronCommand extends Command {
         }
 
         Patron deletePatron = lastShownList.get(targetIndex.getZeroBased());
+        if (model.isBorrowingSomeBook(deletePatron)) {
+            throw new CommandException(MESSAGE_HAS_BORROWED_BOOKS);
+        }
         model.deletePatron(deletePatron);
-        return new CommandResult(String.format(MESSAGE_DELETE_PATRON_SUCCESS, deletePatron));
+        String notification = model.updateBookAfterPatronDelete(deletePatron);
+        if (!notification.isEmpty()) {
+            model.updateFilteredBookList(Model.PREDICATE_SHOW_ALL_BOOKS);
+        }
+        return new CommandResult(String.format(MESSAGE_DELETE_PATRON_SUCCESS, deletePatron) + notification);
     }
 
     @Override
